Random-init mode in main test and 64bit indexing support (#118)
* Implement random-init mode in main test
* Modified buffer indexing to 64-bit to prevent overflow with large sizes

Added random initialization mode for generating random input matrices
and skip validation code for faster performance runs

AMD-Internal: CPUPL-7435

+++ test/main/validate_src/validate_spffrt2.c
@@ -15,7 +15,7 @@ extern double time_min;
 void validate_spffrt2(char *tst_api, integer n, integer ncolm, void *A, void *AP, integer datatype,
                       double err_thresh, void *params)
 {
-    integer i, j, di;
+    size_t i, j, di;
     void *work = NULL;
     void *L = NULL, *D = NULL, *T = NULL;
     double residual = 0.;
